@@ -1,13 +1,15 @@
 import {Component, OnInit, Input} from '@angular/core';
 import {map} from 'rxjs/operators';
 import {Location} from '@angular/common';
-import {Area, ECOE, Round, Schedule, Shift, Stage, Station, Student} from '@models/index';
+import {Area, ECOE, Round, Schedule, Shift, Stage, Station, Student, ApiPermissions, Permission, User} from '@models/index';
 import {Router, ActivatedRoute} from '@angular/router';
 import {TranslateService} from '@ngx-translate/core';
-import {NzMessageService, NzModalService} from 'ng-zorro-antd';
-import {FormGroup, FormBuilder, Validators} from '@angular/forms';
-import {ChronoService} from '@services/chrono/chrono.service';
-import {Item, Pagination} from '@openecoe/potion-client';
+import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+import { ChronoService } from '@services/chrono/chrono.service';
+import { Item, Pagination } from '@openecoe/potion-client';
+import { NzMessageService } from 'ng-zorro-antd/message';
+import { NzModalService } from 'ng-zorro-antd/modal';
+import { ApiService } from '@app/services/api/api.service';
 
 interface ISummaryItems {
   total: number;
@@ -32,6 +34,7 @@ export class EcoeInfoComponent implements OnInit {
   stages: ISummaryItems = {total: 0, show: false, loading: true};
   rounds: ISummaryItems = {total: 0, show: false, loading: true};
   shifts: ISummaryItems = {total: 0, show: false, loading: true};
+  evaluators: ISummaryItems = {total: 0, show: false, loading: true};
 
   show_areas: Boolean = true;
 
@@ -58,7 +61,8 @@ export class EcoeInfoComponent implements OnInit {
     private route: ActivatedRoute,
     private fb: FormBuilder,
     private chronoService: ChronoService,
-    private modalSrv: NzModalService) {
+    private modalSrv: NzModalService,
+    private apiService: ApiService) {
   }
 
   ngOnInit() {
@@ -118,6 +122,13 @@ export class EcoeInfoComponent implements OnInit {
         this.shifts.loading = false;
         this.shifts.show = this.show_planner;
       });
+      
+      this.apiService.getEvaluators(this.ecoe).then((evaluators: User[]) => {
+        this.evaluators.total = evaluators.length;
+        this.evaluators.loading = false;
+      });
+
+  
 
     });
   }
